requirements.txt
pytz
asyncio
dnspython
motor
pyrogram
python-dotenv
tgcrypto